+++ culfw/clib/delay.h
@@ -2,7 +2,7 @@
 #define _MYDELAY_H
 
 extern void my_delay_us( uint16_t d );
-extern void my_delay_ms( uint16_t d );
+extern void my_delay_ms( uint8_t d );
 
 #endif
